pyproject.toml
[tool.poetry]
name = "jasonapi for jasonproject"
version = "0.0.2"
description = "JasonAPI for JasonProject"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
SQLAlchemy = "^2.0.19"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.23.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "^4.0.1"
aiohttp = "^3.8.5"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
